--- app/pages/sections/order-items/catalogue-solutions/edit-solution/commonManifest.json
@@ -1,12 +1,0 @@
-{
-  "backLinkText": "Go back",
-  "title": "information for",
-  "description": "The information we have about this order has been included, but can be changed if required. Refer to the guidance provided if you’re unsure what to enter.",
-  "deleteButton": {
-    "text": "Delete Catalogue Solution"
-  },
-  "editButton": {
-    "text": "Edit Service Recipients"
-  },
-  "saveButtonText": "Save"
-}
