--- app/pages/sections/order-items/catalogue-solutions/edit-solution/controller.js
@@ -1,29 +1,0 @@
-import { getContext } from './contextCreator';
-import commonManifest from './commonManifest.json';
-import { getSelectedPriceManifest } from '../../../../../helpers/controllers/manifestProvider';
-
-export const getOrderItemContext = async ({
-  orderId,
-  orderItemId,
-  orderItemType,
-  solutionName,
-  selectedPrice,
-  formData,
-  recipients,
-}) => {
-  const selectedPriceManifest = getSelectedPriceManifest({
-    orderItemType,
-    provisioningType: selectedPrice.provisioningType,
-    type: selectedPrice.type,
-  });
-
-  return getContext({
-    commonManifest,
-    selectedPriceManifest,
-    orderId,
-    orderItemId,
-    solutionName,
-    formData,
-    recipients,
-  });
-};
