--- app/pages/sections/order-items/catalogue-solutions/edit-solution/template.njk
@@ -1,90 +1,0 @@
-{% extends 'includes/layout.njk' %}
-{% from 'components/back-link/macro.njk' import backLink %}
-{% from 'components/bc-button/macro.njk' import bcButton %}
-{% from 'components/bc-error-summary/macro.njk' import bcErrorSummary %}
-{% from 'components/bc-table/macro.njk' import bcTable %}
-{% from 'components/date-field/macro.njk' import dateField %}
-{% from 'components/text-field/macro.njk' import textField %}
-{% from 'components/radiobutton-options/macro.njk' import radiobuttonOptions %}
-{% from 'components/view/components/view-expandable-section/macro.njk' import viewExpandableSection %}
-
-{% block body %}
-  <div class="nhsuk-width-container" data-test-id="order-item-page">
-    <div data-test-id="go-back-link">
-      {{ backLink({
-        "href": backLinkHref,
-        "text": backLinkText
-      }) }}
-    </div>
-
-    <div class="nhsuk-grid-row">
-      <div class="nhsuk-grid-column-two-thirds">
-        {% if (errors.length > 0) %}
-            {{ bcErrorSummary({
-              errors: errors
-            }) }}
-        {% endif %}
-        <h1 data-test-id="order-item-page-title" class="nhsuk-u-margin-top-7 nhsuk-u-margin-bottom-3">{{ title }}</h1>
-        <h2 data-test-id="order-item-page-description" class="nhsuk-u-font-size-24 nhsuk-u-margin-bottom-6">{{ description }}</h2>
-      </div>
-
-      <form method="post">
-        <div class="nhsuk-grid-column-two-thirds">
-          <input type="hidden" name="_csrf" value="{{ csrfToken }}" />
-
-          <div class="nhsuk-u-margin-bottom-5">
-            <div class="">
-              {{ textField({ 
-                question: questions.price,
-                classes: "nhsuk-input--width-10"
-              }) }}
-              {{ viewExpandableSection(questions.price.expandableSection) }}
-            </div>
-          </div>
-
-        </div>
-
-        <div class="nhsuk-grid-column-full">
-          <hr class="nhsuk-u-margin-top-0 nhsuk-u-margin-bottom-3 bc-c-order-table__hr"/>
-          <div data-test-id="solution-table" class="nhsuk-u-margin-left-5 nhsuk-u-margin-top-0">
-            {{ bcTable({
-              columnInfo: solutionTable.columnInfo,
-              classes: solutionTable.classes,
-              data: solutionTable.items
-            }) }}
-          </div>
-        </div>
-
-        <div class="nhsuk-grid-column-full">
-            <div class="bc-u-inline-block">
-              {{ bcButton({
-                  dataTestId: "delete-button",
-                  text: deleteButton.text,
-                  href: deleteButton.href,
-                  disabled: deleteButton.disabled,
-                  classes: "nhsuk-u-margin-top-9 nhsuk-button--secondary nhsuk-u-margin-right-5"
-              }) }}
-            </div>
-          <div class="bc-u-float-right">
-            <div class="bc-u-inline-block">
-              {{ bcButton({
-                  dataTestId: "edit-button",
-                  text: editButton.text,
-                  href: editButton.href,
-                  disabled: editButton.disabled,
-                  classes: "nhsuk-u-margin-top-9 nhsuk-button--secondary nhsuk-u-margin-right-5"
-              }) }}
-            </div>
-            <div class="bc-u-inline-block">
-              {{ bcButton({
-                  dataTestId: "save-button",
-                  text: saveButtonText,
-                  classes: "nhsuk-u-margin-top-9"
-              }) }}
-            </div>
-          </div>
-        </div>
-      </form>
-    </div>
-  </div>
-{% endblock %}
