+++ app/pages/sections/order-items/catalogue-solutions/order-item/controller.js
@@ -1,31 +1,15 @@
-import { getContext, getErrorContext } from './contextCreator';
+import { getContext } from './contextCreator';
 import commonManifest from './commonManifest.json';
 import { getSelectedPriceManifest } from '../../../../../helpers/controllers/manifestProvider';
-
-const formatFormData = ({ formData }) => ({
-  'deliveryDate-day': formData['deliveryDate-day']
-    ? formData['deliveryDate-day'].trim() : undefined,
-  'deliveryDate-month': formData['deliveryDate-month']
-    ? formData['deliveryDate-month'].trim() : undefined,
-  'deliveryDate-year': formData['deliveryDate-year']
-    ? formData['deliveryDate-year'].trim() : undefined,
-  quantity: formData.quantity
-    ? formData.quantity.trim() : undefined,
-  price: formData.price && formData.price.length > 0
-    ? formData.price.trim() : undefined,
-  selectEstimationPeriod: formData.selectEstimationPeriod
-    ? formData.selectEstimationPeriod.trim() : undefined,
-});
 
 export const getOrderItemContext = async ({
   orderId,
   orderItemId,
   orderItemType,
   solutionName,
-  odsCode,
-  serviceRecipientName,
   selectedPrice,
   formData,
+  recipients,
 }) => {
   const selectedPriceManifest = getSelectedPriceManifest({
     orderItemType,
@@ -39,30 +23,7 @@ export const getOrderItemContext = async ({
     orderId,
     orderItemId,
     solutionName,
-    serviceRecipientName,
-    odsCode,
-    selectedPrice,
     formData,
+    recipients,
   });
-};
-
-export const getOrderItemErrorPageContext = (params) => {
-  const formattedData = formatFormData({
-    formData: params.formData,
-  });
-
-  const selectedPriceManifest = getSelectedPriceManifest({
-    orderItemType: params.orderItemType,
-    provisioningType: params.selectedPrice.provisioningType,
-    type: params.selectedPrice.type,
-  });
-
-  const updatedParams = {
-    ...params,
-    commonManifest,
-    selectedPriceManifest,
-    formData: formattedData,
-  };
-
-  return getErrorContext(updatedParams);
 };
