+++ app/pages/sections/order-items/catalogue-solutions/order-item/controller.test.js
@@ -1,6 +1,5 @@
 import {
   getOrderItemContext,
-  getOrderItemErrorPageContext,
 } from './controller';
 import * as contextCreator from './contextCreator';
 import * as getSelectedPriceManifest from '../../../../../helpers/controllers/manifestProvider';
@@ -39,7 +38,6 @@ describe('catalogue-solutions order-item controller', () => {
         orderItemType,
         solutionName: 'solution-name',
         selectedRecipientId: 'fake-recipient-id',
-        serviceRecipientName: 'Some service recipient 1',
         selectedPriceId: 'some-price-id',
         selectedPrice,
       });
@@ -49,33 +47,6 @@ describe('catalogue-solutions order-item controller', () => {
         orderItemType,
         provisioningType: selectedPrice.provisioningType,
         type: selectedPrice.type,
-      });
-    });
-
-    it('should call getContext with the correct params', async () => {
-      const selectedPriceManifest = { description: 'fake manifest' };
-      getSelectedPriceManifest.getSelectedPriceManifest.mockReturnValue(selectedPriceManifest);
-
-      await getOrderItemContext({
-        orderId: 'order-1',
-        solutionName: 'solution-name',
-        odsCode: 'fake-recipient-id',
-        serviceRecipientName: 'Some service recipient 1',
-        selectedPriceId: 'some-price-id',
-        selectedPrice,
-        formData: { price: 0.1 },
-      });
-
-      expect(contextCreator.getContext.mock.calls.length).toEqual(1);
-      expect(contextCreator.getContext).toHaveBeenCalledWith({
-        commonManifest: { title: 'fake manifest' },
-        selectedPriceManifest,
-        odsCode: 'fake-recipient-id',
-        orderId: 'order-1',
-        serviceRecipientName: 'Some service recipient 1',
-        solutionName: 'solution-name',
-        selectedPrice,
-        formData: { price: 0.1 },
       });
     });
 
@@ -84,89 +55,34 @@ describe('catalogue-solutions order-item controller', () => {
       getSelectedPriceManifest.getSelectedPriceManifest.mockReturnValue(selectedPriceManifest);
 
       const formData = {
-        'deliveryDate-year': '2020',
-        'deliveryDate-month': '04',
-        'deliveryDate-day': '27',
-        quantity: 3,
-        selectEstimationPeriod: 'month',
+        deliveryDate: {
+          day: 9,
+          month: 2,
+          year: 2021,
+        },
         price: 0.1,
       };
+      const recipients = [{}];
 
       await getOrderItemContext({
         orderId: 'order-1',
         solutionName: 'solution-name',
-        odsCode: 'fake-recipient-id',
-        serviceRecipientName: 'Some service recipient 1',
         selectedPriceId: 'some-price-id',
+        orderItemId: 'order-item-id',
         selectedPrice,
         formData,
+        recipients,
       });
 
       expect(contextCreator.getContext.mock.calls.length).toEqual(1);
       expect(contextCreator.getContext).toHaveBeenCalledWith({
         commonManifest: { title: 'fake manifest' },
         selectedPriceManifest,
-        odsCode: 'fake-recipient-id',
         orderId: 'order-1',
-        serviceRecipientName: 'Some service recipient 1',
         solutionName: 'solution-name',
-        selectedPrice,
+        orderItemId: 'order-item-id',
         formData,
-      });
-    });
-  });
-
-  describe('getOrderItemErrorPageContext', () => {
-    afterEach(() => {
-      jest.resetAllMocks();
-    });
-
-    it('should call getSelectedPriceManifest with the correct params', async () => {
-      await getOrderItemErrorPageContext({
-        orderId: 'order-1',
-        orderItemType,
-        itemName: 'item-name',
-        selectedPriceId: 'some-price-id',
-        selectedPrice,
-        formData: { quantity: '100' },
-        validationErrors: {},
-      });
-
-      expect(getSelectedPriceManifest.getSelectedPriceManifest.mock.calls.length).toEqual(1);
-      expect(getSelectedPriceManifest.getSelectedPriceManifest).toHaveBeenCalledWith({
-        orderItemType,
-        provisioningType: selectedPrice.provisioningType,
-        type: selectedPrice.type,
-      });
-    });
-
-    it('should call getOrderItemErrorPageContext with the correct params', async () => {
-      const selectedPriceManifest = { description: 'fake manifest' };
-      getSelectedPriceManifest.getSelectedPriceManifest.mockReturnValue(selectedPriceManifest);
-
-      await getOrderItemErrorPageContext({
-        orderId: 'order-1',
-        itemName: 'item-name',
-        selectedPriceId: 'some-price-id',
-        selectedPrice,
-        formData: {
-          quantity: '100  ',
-          price: '  0.1',
-        },
-      });
-
-      expect(contextCreator.getErrorContext.mock.calls.length).toEqual(1);
-      expect(contextCreator.getErrorContext).toHaveBeenCalledWith({
-        commonManifest: { title: 'fake manifest' },
-        selectedPriceManifest,
-        orderId: 'order-1',
-        itemName: 'item-name',
-        selectedPriceId: 'some-price-id',
-        selectedPrice,
-        formData: {
-          quantity: '100',
-          price: '0.1',
-        },
+        recipients,
       });
     });
   });
