--- app/pages/sections/order-items/catalogue-solutions/order-item/flat/ondemand/manifest.json
@@ -1,92 +1,0 @@
-{
-  "questions": {
-    "deliveryDate": {
-      "id": "deliveryDate",
-      "type": "date",
-      "mainAdvice": "What is the planned delivery date?",
-      "additionalAdvice": "For example 14 01 2020"
-    },
-    "quantity": {
-      "id": "quantity",
-      "mainAdvice": "How many do you think you'll need?",
-      "rows": 3,
-      "expandableSection": {
-        "dataTestId": "view-section-quantity-id",
-        "title": "What amount should I enter?",
-        "innerComponent": "Estimate the quantity you think you'll need either per month or per year."
-      }
-    }, 
-    "selectEstimationPeriod": {
-      "id": "selectEstimationPeriod",
-      "type": "radio",
-      "mainAdvice": "Over what period did you estimate the quantity?",
-      "options": [
-        {
-          "value": "month",
-          "text": "Per month"
-        },
-        {
-          "value": "year",
-          "text": "Per year"
-        }
-      ],
-      "expandableSection": {
-        "dataTestId": "view-section-estimation-period-id",
-        "title": "What period should I enter?",
-        "innerComponent": "This should be based on how you estimated the quantity you want to order."
-      }
-    }
-  },
-  "addPriceTable": {
-    "columnInfo": [
-      {
-        "data": "What price have you agreed? (£)",
-        "width": "50%",
-        "classes": "nhsuk-u-font-size-16"
-      },
-      {
-        "data": "Unit of order",
-        "width": "50%",
-        "classes": "nhsuk-u-font-size-16"
-      }
-    ],
-    "cellInfo": {
-      "price": {
-        "question": {
-          "type": "input",
-          "id": "price"
-        },
-        "classes": "nhsuk-input--width-10",
-        "expandableSection": {
-          "dataTestId": "view-section-input-id",
-          "title": "What price should I enter?",
-          "innerComponent": "You can change the list price if you've agreed a different rate with the supplier."
-        }
-      },
-      "unitOfOrder": {
-        "dataTestId": "unit-of-order"
-      }
-    }
-  },
-  "errorMessages": {
-    "QuantityRequired": "Enter an estimated quantity",
-    "QuantityMustBeANumber": "Estimated quantity must be a number",
-    "QuantityInvalid": "Estimated quantity must be a whole number",
-    "QuantityGreaterThanZero": "Estimated quantity must be greater than zero",
-    "QuantityLessThanMax": "Maximum quantity exceeded, enter a lower estimated quantity",
-    "EstimationPeriodRequired": "Select an estimation period",
-    "PriceRequired": "Enter a price",
-    "PriceMustBeANumber": "Price must be a number",
-    "PriceMoreThan3dp": "Price must be to a maximum of 3 decimal places",
-    "PriceGreaterThanOrEqualToZero": "Price cannot be negative",
-    "PriceLessThanMax": "Maximum price exceeded, enter a lower price",
-    "DeliveryDateRequired": "Enter a planned delivery date",
-    "DeliveryDateInvalid": "Enter a planned delivery date in a valid format",
-    "DeliveryDateOutsideDeliveryWindow": "Planned delivery date must be within 42 months from the commencement date for this Call-off Agreement",
-    "DeliveryDateNotReal": "Planned delivery date must be a real date",
-    "DeliveryDateDayRequired": "Planned delivery date must include a day",
-    "DeliveryDateMonthRequired": "Planned delivery date must include a month",
-    "DeliveryDateYearRequired": "Planned delivery date must include a year",
-    "DeliveryDateYearLength": "Year must be four numbers, for example 2020"
-  }
-}
