+++ app/pages/sections/order-items/catalogue-solutions/order-item/flat/patient/manifest.json
@@ -1,49 +1,61 @@
 {
   "questions": {
-    "deliveryDate": {
-      "id": "deliveryDate",
-      "type": "date",
-      "mainAdvice": "What is the planned delivery date?",
-      "additionalAdvice": "For example 14 01 2020"
-    },
-    "quantity": {
-      "id": "quantity",
-      "mainAdvice": "What is your practice list size?",
+    "price": {
+      "id": "price",
+      "mainAdvice": "What price have you agreed? (£)",
       "expandableSection": {
-        "dataTestId": "view-section-quantity-id",
-        "title": "What list size should I enter?",
-        "innerComponent": "This figure is usually the number of patients registered with the Service Recipient. If you're unsure how this relates to your order you can contact the helpdesk using the link at the bottom of the page."
+        "dataTestId": "view-section-price-id",
+        "title": "What price should I enter?",
+        "innerComponent": "Enter the price you've agreed with the supplier. We've included the list price, but this can be changed if required."
       }
     }
   },
-  "addPriceTable": {
+  "solutionTable": {
     "columnInfo": [
       {
-        "data": "What price have you agreed? (£)",
-        "width": "50%",
-        "classes": "nhsuk-u-font-size-16"
+        "data": "Recipient name (ODS code)",
+        "width": "33%",
+        "classes": " nhsuk-u-padding-bottom-5 nhsuk-u-font-size-20"
       },
       {
-        "data": "Unit of order",
-        "width": "50%",
-        "classes": "nhsuk-u-font-size-16"
+        "data": "Practice list size",
+        "width": "33%",
+        "classes": "nhsuk-u-padding-bottom-5 nhsuk-u-font-size-20"
+      },
+      {
+        "data": "Planned delivery date",
+        "width": "33%",
+        "classes": "nhsuk-u-padding-bottom-3 nhsuk-u-font-size-20",
+        "additionalAdvice": "For example 14 01 2020"
       }
     ],
     "cellInfo": {
-      "price": {
+      "recipient": {
+        "dataTestId": "recipient",
+        "classes": "nhsuk-u-padding-top-5"
+      },
+      "practiceSize": {
         "question": {
           "type": "input",
-          "id": "price"
+          "id": "practiceSize"
         },
-        "classes": "nhsuk-input--width-10",
         "expandableSection": {
-          "dataTestId": "view-section-input-id",
-          "title": "What price should I enter?",
-          "innerComponent": "You can change the list price if you've agreed a different rate with the supplier."
-        }
+          "dataTestId": "view-section-input-id-practice",
+          "title": "What list size should I enter?",
+          "innerComponent": "Enter the amount you wish to order. This is usually based on each Service Recipient's practice list size to help calculate an estimated price, but the figure can be changed if required.<br><br>As you’re ordering per patient, we've included each practice list size if we have it. If it's not included, you'll need to add it yourself."
+        },
+        "classes": "nhsuk-u-margin-top-5 nhsuk-input--width-10"
       },
-      "unitOfOrder": {
-        "dataTestId": "unit-of-order"
+      "plannedDeliveryDate": {
+        "question": {
+          "type": "date",
+          "id": "plannedDeliveryDate"
+        },
+        "expandableSection": {
+          "dataTestId": "view-section-input-id-date",
+          "title": "What date should I enter?",
+          "innerComponent": "Enter the planned delivery date you've agreed with the supplier. This is the day that each of the listed Service Recipients will start using this Catalogue Solution.<br><br>We've included the commencement date for this Call-off Agreement, but you can change it if you’ve agreed a different delivery date."
+        }
       }
     }
   },
