+++ app/pages/sections/order-items/catalogue-solutions/order-item/flat/patient/ui-tests/general.ui.test.js
@@ -8,8 +8,6 @@ import { sessionKeys } from '../../../../../../../../helpers/routes/sessionHelpe
 
 const pageUrl = 'http://localhost:1234/order/organisation/order-id/catalogue-solutions/neworderitem';
 
-const getLocation = ClientFunction(() => document.location.href);
-
 const selectedPrice = {
   priceId: 2,
   provisioningType: 'patient',
@@ -31,6 +29,9 @@ const itemNameInSession = 'solution-name';
 const selectedRecipientIdInSession = 'recipient-1';
 const selectedRecipientNameInSession = 'recipient-name';
 const selectedPriceIdInSession = 'price-1';
+const catalogueSolutionIdInSession = 'solution-1';
+const plannedDeliveryDateInSession = '2020-10-10';
+const recipientsInSession = JSON.stringify([{ name: 'recipient-name', odsCode: 'code' }]);
 
 const orderItemPageDataInSession = JSON.stringify({
   itemId: itemIdInSession,
@@ -61,6 +62,9 @@ const pageSetup = async (setup = defaultPageSetup) => {
     await setState(ClientFunction)(sessionKeys.selectedItemId, itemIdInSession);
     await setState(ClientFunction)(sessionKeys.selectedItemName, itemNameInSession);
     await setState(ClientFunction)(sessionKeys.selectedPriceId, selectedPriceIdInSession);
+    await setState(ClientFunction)(sessionKeys.catalogueSolutionId, catalogueSolutionIdInSession);
+    await setState(ClientFunction)(sessionKeys.plannedDeliveryDate, plannedDeliveryDateInSession);
+    await setState(ClientFunction)(sessionKeys.recipients, recipientsInSession);
   }
   if (setup.postRoute) {
     await setState(ClientFunction)(sessionKeys.orderItemPageData, orderItemPageDataInSession);
@@ -73,208 +77,75 @@ fixture('Catalogue-solutions - flat patient - general')
     await nockAndErrorCheck(nock, t);
   });
 
-test('should render a text field for the quantity question', async (t) => {
+test('should render the solution table headings', async (t) => {
   await pageSetup();
   await t.navigateTo(pageUrl);
 
-  const quantity = Selector('[data-test-id="question-quantity"]');
-  const quantityLabel = quantity.find('label.nhsuk-label');
+  const solutionTable = Selector('div[data-test-id="solution-table"]');
+  const solutionNameColumnHeading = solutionTable.find('[data-test-id="column-heading-0"]');
+  const practiceSizeColumnHeading = solutionTable.find('[data-test-id="column-heading-1"]');
+  const dateColumnHeading = solutionTable.find('[data-test-id="column-heading-2"]');
 
   await t
-    .expect(quantity.exists).ok()
-    .expect(await extractInnerText(quantityLabel)).eql(content.questions.quantity.mainAdvice)
-    .expect(quantity.find('input').count).eql(1);
-});
-
-test('should render an expandable section for the quantity question', async (t) => {
-  await pageSetup();
-  await t.navigateTo(pageUrl);
-
-  const expandableSection = Selector('[data-test-id="view-section-quantity-id"]');
-
-  await t
-    .expect(expandableSection.exists).ok()
-    .expect(await extractInnerText(expandableSection)).eql(content.questions.quantity.expandableSection.title)
-    .expect(expandableSection.find('details[open]').exists).notOk()
-    .click(expandableSection.find('summary'))
-    .expect(expandableSection.find('details[open]').exists).ok()
-    .expect(await extractInnerText(expandableSection.find('.nhsuk-details__text')))
-    .eql(content.questions.quantity.expandableSection.innerComponent);
-});
-
-test('should not render a selectEstimationPeriod question for flat patient', async (t) => {
-  await pageSetup();
-  await t.navigateTo(pageUrl);
-
-  const selectEstimationPeriodRadioOptions = Selector('[data-test-id="question-selectEstimationPeriod"]');
-
-  await t
-    .expect(selectEstimationPeriodRadioOptions.exists).notOk();
-});
-
-test('should render the price table headings', async (t) => {
-  await pageSetup();
-  await t.navigateTo(pageUrl);
-
-  const priceTable = Selector('div[data-test-id="price-table"]');
-  const priceColumnHeading = priceTable.find('[data-test-id="column-heading-0"]');
-  const unitColumnHeading = priceTable.find('[data-test-id="column-heading-1"]');
-
-  await t
-    .expect(priceTable.exists).ok()
-    .expect(priceColumnHeading.exists).ok()
-    .expect(await extractInnerText(priceColumnHeading)).eql(content.addPriceTable.columnInfo[0].data)
-    .expect(unitColumnHeading.exists).ok()
-    .expect(await extractInnerText(unitColumnHeading)).eql(content.addPriceTable.columnInfo[1].data);
+    .expect(solutionTable.exists).ok()
+    .expect(solutionNameColumnHeading.exists).ok()
+    .expect(await extractInnerText(solutionNameColumnHeading)).eql(content.solutionTable.columnInfo[0].data)
+    .expect(practiceSizeColumnHeading.exists).ok()
+    .expect(await extractInnerText(practiceSizeColumnHeading)).eql(content.solutionTable.columnInfo[1].data)
+    .expect(dateColumnHeading.exists).ok()
+    .expect(await extractInnerText(dateColumnHeading)).eql(`${content.solutionTable.columnInfo[2].data}\n${content.solutionTable.columnInfo[2].additionalAdvice}`);
 });
 
 test('should render the price table content', async (t) => {
   await pageSetup();
   await t.navigateTo(pageUrl);
 
-  const table = Selector('div[data-test-id="price-table"]');
+  const table = Selector('div[data-test-id="solution-table"]');
   const row = table.find('[data-test-id="table-row-0"]');
-  const priceInput = row.find('[data-test-id="question-price"] input');
-  const expandableSection = row.find('[data-test-id="view-section-input-id"]');
-  const orderUnit = row.find('div[data-test-id="unit-of-order"]');
+  const solutionName = row.find('div[data-test-id="recipient-name-code-recipient"]');
+  const practiceSizeInput = row.find('[data-test-id="question-practiceSize"] input');
+  const practiceSizeExpandableSection = row.find('[data-test-id="view-section-input-id-practice"]');
+  const dateInput = row.find('[data-test-id="question-plannedDeliveryDate"] input');
+  const dayInput = dateInput.nth(0);
+  const monthInput = dateInput.nth(1);
+  const yearInput = dateInput.nth(2);
+  const dateExpandableSection = row.find('[data-test-id="view-section-input-id-date"]');
 
   await t
     .expect(row.exists).ok()
-    .expect(priceInput.exists).ok()
-    .expect(priceInput.getAttribute('value')).eql(selectedPrice.price.toString())
-    .expect(expandableSection.exists).ok()
-    .expect(await extractInnerText(expandableSection)).eql(content.addPriceTable.cellInfo.price.expandableSection.title)
-    .expect(expandableSection.find('details[open]').exists).notOk()
-    .click(expandableSection.find('summary'))
-    .expect(expandableSection.find('details[open]').exists).ok()
-    .expect(await extractInnerText(expandableSection.find('.nhsuk-details__text')))
-    .eql(content.addPriceTable.cellInfo.price.expandableSection.innerComponent)
-    .expect(orderUnit.exists).ok()
-    .expect(await extractInnerText(orderUnit)).eql(`${selectedPrice.itemUnit.description} ${selectedPrice.timeUnit.description}`);
-});
+    .expect(solutionName.exists).ok()
+    .expect(await extractInnerText(solutionName)).eql('recipient-name (code)')
 
-test('should render the price with a value of 0 when returned from the API', async (t) => {
-  const mockSelectedPrice = { ...selectedPrice, price: 0 };
-  await pageSetup({ ...defaultPageSetup, postRoute: true, mockData: mockSelectedPrice });
-  await t.navigateTo(pageUrl);
+    .expect(practiceSizeInput.exists).ok()
+    .expect(practiceSizeExpandableSection.exists).ok()
+    .expect(await extractInnerText(practiceSizeExpandableSection)).eql(content.solutionTable.cellInfo.practiceSize.expandableSection.title)
+    .expect(practiceSizeExpandableSection.find('details[open]').exists).notOk()
+    .click(practiceSizeExpandableSection.find('summary'))
+    .expect(practiceSizeExpandableSection.find('details[open]').exists).ok()
+    .expect(await extractInnerText(practiceSizeExpandableSection.find('.nhsuk-details__text')))
+    .eql(content.solutionTable.cellInfo.practiceSize.expandableSection.innerComponent.replace('<br><br>', ''))
 
-  const table = Selector('div[data-test-id="price-table"]');
-  const row = table.find('[data-test-id="table-row-0"]');
-  const priceInput = row.find('[data-test-id="question-price"] input');
+    .expect(dateInput.exists).ok()
+    .expect(dayInput.getAttribute('id')).eql('plannedDeliveryDate-day')
+    .expect(dayInput.getAttribute('name')).eql('plannedDeliveryDate-day')
+    .expect(dayInput.getAttribute('type')).eql('number')
+    .expect(dayInput.getAttribute('value')).eql('10')
 
-  await t
-    .expect(priceInput.getAttribute('value')).eql(mockSelectedPrice.price.toString());
-});
+    .expect(monthInput.getAttribute('id')).eql('plannedDeliveryDate-month')
+    .expect(monthInput.getAttribute('name')).eql('plannedDeliveryDate-month')
+    .expect(monthInput.getAttribute('type')).eql('number')
+    .expect(monthInput.getAttribute('value')).eql('10')
 
-test('should render select quantity field as errors with error message when no quantity entered causing validation error', async (t) => {
-  await pageSetup({ ...defaultPageSetup, postRoute: true });
-  await t.navigateTo(pageUrl);
+    .expect(yearInput.getAttribute('id')).eql('plannedDeliveryDate-year')
+    .expect(yearInput.getAttribute('name')).eql('plannedDeliveryDate-year')
+    .expect(yearInput.getAttribute('type')).eql('number')
+    .expect(yearInput.getAttribute('value')).eql('2020')
 
-  const orderItemPage = Selector('[data-test-id="order-item-page"]');
-  const saveButton = Selector('[data-test-id="save-button"] button');
-  const quantityField = orderItemPage.find('[data-test-id="question-quantity"]');
-
-  await t
-    .expect(saveButton.exists).ok()
-    .expect(quantityField.find('[data-test-id="text-field-input-error"]').exists).notOk()
-    .click(saveButton);
-
-  await t
-    .expect(quantityField.find('[data-test-id="text-field-input-error"]').exists).ok()
-    .expect(await extractInnerText(quantityField.find('#quantity-error'))).contains(content.errorMessages.QuantityRequired);
-});
-
-test('should render select price field as errors with error message when no price entered causing validation error', async (t) => {
-  await pageSetup({ ...defaultPageSetup, postRoute: true });
-  await t.navigateTo(pageUrl);
-
-  const orderItemPage = Selector('[data-test-id="order-item-page"]');
-  const saveButton = Selector('[data-test-id="save-button"] button');
-  const priceField = orderItemPage.find('[data-test-id="question-price"]');
-
-  await t
-    .expect(priceField.find('[data-test-id="text-field-input-error"]').exists).notOk()
-    .selectText(priceField.find('input')).pressKey('delete')
-    .click(saveButton);
-
-  await t
-    .expect(priceField.find('[data-test-id="text-field-input-error"]').exists).ok()
-    .expect(await extractInnerText(priceField.find('#price-error'))).contains(content.errorMessages.PriceRequired);
-});
-
-test('should anchor to the quantity field when clicking on the quantity required error link in errorSummary ', async (t) => {
-  await pageSetup({ ...defaultPageSetup, postRoute: true });
-  await t.navigateTo(pageUrl);
-
-  const continueButton = Selector('[data-test-id="save-button"] button');
-  const errorSummary = Selector('[data-test-id="error-summary"]');
-
-  await t
-    .expect(errorSummary.exists).notOk()
-    .click(continueButton);
-
-  await t
-    .expect(errorSummary.exists).ok()
-
-    .click(errorSummary.find('li a').nth(1))
-    .expect(getLocation()).eql(`${pageUrl}#quantity`);
-});
-
-test('should anchor to the quantity field when clicking on the numerical quantity error link in errorSummary ', async (t) => {
-  await pageSetup({ ...defaultPageSetup, postRoute: true });
-  await t.navigateTo(pageUrl);
-
-  const continueButton = Selector('[data-test-id="save-button"] button');
-  const errorSummary = Selector('[data-test-id="error-summary"]');
-  const quantity = Selector('[data-test-id="question-quantity"]');
-
-  await t
-    .expect(errorSummary.exists).notOk()
-    .typeText(quantity, 'blah', { paste: true })
-    .click(continueButton);
-
-  await t
-    .expect(errorSummary.exists).ok()
-
-    .click(errorSummary.find('li a').nth(1))
-    .expect(getLocation()).eql(`${pageUrl}#quantity`);
-});
-
-test('should anchor to the price field when clicking on the price required error link in errorSummary ', async (t) => {
-  await pageSetup({ ...defaultPageSetup, postRoute: true });
-  await t.navigateTo(pageUrl);
-
-  const saveButton = Selector('[data-test-id="save-button"] button');
-  const errorSummary = Selector('[data-test-id="error-summary"]');
-  const priceInput = Selector('[data-test-id="question-price"] input');
-
-  await t
-    .expect(errorSummary.exists).notOk()
-    .selectText(priceInput).pressKey('delete')
-    .click(saveButton);
-
-  await t
-    .expect(errorSummary.exists).ok()
-    .click(errorSummary.find('li a').nth(2))
-    .expect(getLocation()).eql(`${pageUrl}#price`);
-});
-
-test('should anchor to the price field when clicking on the numerical price error link in errorSummary ', async (t) => {
-  await pageSetup({ ...defaultPageSetup, postRoute: true });
-  await t.navigateTo(pageUrl);
-
-  const continueButton = Selector('[data-test-id="save-button"] button');
-  const errorSummary = Selector('[data-test-id="error-summary"]');
-  const priceInput = Selector('[data-test-id="question-price"] input');
-
-  await t
-    .expect(errorSummary.exists).notOk()
-    .typeText(priceInput, 'blah', { paste: true })
-    .click(continueButton);
-
-  await t
-    .expect(errorSummary.exists).ok()
-
-    .click(errorSummary.find('li a').nth(2))
-    .expect(getLocation()).eql(`${pageUrl}#price`);
+    .expect(dateExpandableSection.exists).ok()
+    .expect(await extractInnerText(dateExpandableSection)).eql(content.solutionTable.cellInfo.plannedDeliveryDate.expandableSection.title)
+    .expect(dateExpandableSection.find('details[open]').exists).notOk()
+    .click(dateExpandableSection.find('summary'))
+    .expect(dateExpandableSection.find('details[open]').exists).ok()
+    .expect(await extractInnerText(dateExpandableSection.find('.nhsuk-details__text')))
+    .eql(content.solutionTable.cellInfo.plannedDeliveryDate.expandableSection.innerComponent.replace('<br><br>', ''));
 });
