+++ app/pages/sections/order-items/catalogue-solutions/routes.js
@@ -8,7 +8,7 @@ import {
 import { catalogueSolutionsSelectRoutes } from './select/routes';
 import {
   getOrderItemContext,
-} from './edit-solution/controller';
+} from './order-item/controller';
 // import { validateOrderItemForm } from '../../../../helpers/controllers/validateOrderItemForm';
 import { getOrderItemPageData } from '../../../../helpers/routes/getOrderItemPageData';
 // import { saveOrderItem } from '../../../../helpers/controllers/saveOrderItem';
@@ -72,7 +72,7 @@ export const catalogueSolutionsRoutes = (authProvider, addContext, sessionManage
     });
 
     logger.info(`navigating to order ${orderId} catalogue-solutions order item page`);
-    return res.render('pages/sections/order-items/catalogue-solutions/edit-solution/template.njk', addContext({ context, user: req.user, csrfToken: req.csrfToken() }));
+    return res.render('pages/sections/order-items/catalogue-solutions/order-item/template.njk', addContext({ context, user: req.user, csrfToken: req.csrfToken() }));
   }));
 
   // router.post('/:orderItemId', authProvider.authorise({ claim: 'ordering' }),
